refactor: merge checkpoint modules + extract shared execution helpers
- Merge session-checkpoint.ts into checkpoint.ts (single module for
  step-level restart-resume and session-level rollback snapshots)
- Remove dead code: loadCheckpoint, saveFailedSnapshot, clearSessionCheckpoints
- Extract 3 shared helpers into helpers/run-agent.ts:
  createBaseEventSender, loadMemoryForPrompt, cleanupAgentRun
- Deduplicate event dispatch, memory loading, and cleanup logic
  between server.ts resumeFromCheckpoint and routes/agent.ts POST handler
