+++ agent/core/checkpoint.ts
@@ -1,31 +1,38 @@
 /**
- * Checkpoint — Agent 运行检查点持久化，支持服务重启后恢复未完成的任务
+ * Checkpoint — Agent 状态持久化：重启恢复 + 会话回滚
  *
- * 每次 Agent 执行一步后保存当前状态（history, step, task, model 等）到 JSON 文件。
- * 服务重启时扫描 checkpoints 目录，恢复所有未完成的运行。
+ * 两套机制共用一个模块：
+ *   1. Step 级检查点 — 每步覆盖写入，用于服务重启后恢复未完成任务
+ *   2. Session 级快照 — 每步独立文件，用于运行中回滚到任意历史步骤
  *
- * 存储位置：{CHECKPOINT_DIR}/checkpoints/{runId}.json
- * 使用原子写入（先写 .tmp 再 rename）防止写入中途崩溃导致文件损坏。
+ * 存储位置：
+ *   Step 级：{dir}/checkpoints/{runId}.json          （单文件，覆盖写入）
+ *   Session 级：{dir}/session-checkpoints/{runId}/session-healthy-{step}.json
  *
- * 调用场景：
- *   - agent/desktop/agent.js 的 onCheckpoint 回调：每步执行后 saveCheckpoint
- *   - server.js 启动时：listCheckpoints 发现未完成任务 → resumeFromCheckpoint 恢复
- *   - routes/agent.js finally 块：任务完成后 removeCheckpoint 清理
+ * 所有写入使用原子操作（先写 .tmp 再 rename）防崩溃损坏。
  */
 
-import { mkdir, writeFile, readFile, unlink, readdir, rename } from 'node:fs/promises';
+import { mkdir, writeFile, readFile, unlink, readdir, rename, rm } from 'node:fs/promises';
 import { join } from 'node:path';
 import { log } from '../../helpers/logger.ts';
 
-function checkpointPath(dir, runId) {
+// ─── 常量 ────────────────────────────────────────────────
+
+const HEALTH_CHECKPOINT_INTERVAL = 1;
+const KEEP_HEALTHY = 30;
+const KEEP_FAILED = 3;
+
+// ─── Step 级检查点（重启恢复）────────────────────────────
+
+function checkpointPath(dir: string, runId: string) {
   return join(dir, 'checkpoints', `${runId}.json`);
 }
 
-function tmpPath(dir, runId) {
+function tmpPath(dir: string, runId: string) {
   return join(dir, 'checkpoints', `${runId}.json.tmp`);
 }
 
-export async function saveCheckpoint(dir, data) {
+export async function saveCheckpoint(dir: string, data: any) {
   const cpDir = join(dir, 'checkpoints');
   await mkdir(cpDir, { recursive: true });
   const tmp = tmpPath(dir, data.runId);
@@ -34,17 +41,7 @@ export async function saveCheckpoint(dir, data) {
   await rename(tmp, dest);
 }
 
-export async function loadCheckpoint(dir, runId) {
-  try {
-    const raw = await readFile(checkpointPath(dir, runId), 'utf8');
-    return JSON.parse(raw);
-  } catch (err) {
-    log.warn(`Checkpoint load failed for ${runId}:`, err.message);
-    return null;
-  }
-}
-
-export async function listCheckpoints(dir) {
+export async function listCheckpoints(dir: string) {
   const cpDir = join(dir, 'checkpoints');
   try {
     const files = await readdir(cpDir);
@@ -54,7 +51,7 @@ export async function listCheckpoints(dir) {
       try {
         const raw = await readFile(join(cpDir, f), 'utf8');
         checkpoints.push(JSON.parse(raw));
-      } catch (err) {
+      } catch (err: any) {
         log.warn(`Checkpoint parse failed: ${f}:`, err.message);
       }
     }
@@ -64,15 +61,15 @@ export async function listCheckpoints(dir) {
   }
 }
 
-export async function removeCheckpoint(dir, runId) {
+export async function removeCheckpoint(dir: string, runId: string) {
   try {
     await unlink(checkpointPath(dir, runId));
-  } catch (err) {
+  } catch (err: any) {
     log.debug(`Checkpoint remove failed for ${runId}:`, err.message);
   }
 }
 
-export async function clearCheckpoints(dir) {
+export async function clearCheckpoints(dir: string) {
   const cpDir = join(dir, 'checkpoints');
   try {
     const files = await readdir(cpDir);
@@ -81,7 +78,146 @@ export async function clearCheckpoints(dir) {
         await unlink(join(cpDir, f)).catch(() => {});
       }
     }
-  } catch (err) {
+  } catch (err: any) {
     log.debug('Clear checkpoints failed:', err.message);
   }
 }
+
+// ─── Session 级快照（回滚）─────────────────────────────────
+
+function sessionDir(dir: string, runId: string) {
+  return join(dir, 'session-checkpoints', runId);
+}
+
+function healthyPath(dir: string, runId: string, step: number) {
+  return join(sessionDir(dir, runId), `session-healthy-${step}.json`);
+}
+
+function failedPath(dir: string, runId: string, step: number) {
+  return join(sessionDir(dir, runId), `session-failed-${step}.json`);
+}
+
+function assessHealth(result: any) {
+  if (!result || typeof result !== 'string') return 'unknown';
+  if (result.startsWith('执行失败')) return 'unhealthy';
+  if (result.startsWith('操作未获批准')) return 'degraded';
+  return 'healthy';
+}
+
+function sanitizeState(state: any) {
+  if (!state) return null;
+  const safe = { ...state };
+  delete safe.chromium;
+  delete safe.browserCandidatePaths;
+  delete safe.onEvent;
+  delete safe.browserSession;
+  delete safe.observeDesktop;
+  safe.browserSessionActive = Boolean(state.browserSession);
+  return safe;
+}
+
+async function pruneSnapshots(dir: string, runId: string, type: string, keepCount: number) {
+  const cpDir = sessionDir(dir, runId);
+  try {
+    const files = await readdir(cpDir);
+    for (const f of files) {
+      if (f.endsWith('.tmp')) {
+        await unlink(join(cpDir, f)).catch(() => {});
+      }
+    }
+    const snapshots = files
+      .filter(f => f.startsWith(`session-${type}-`) && f.endsWith('.json'))
+      .map(f => {
+        const match = f.match(new RegExp(`session-${type}-(\\d+)\\.json$`));
+        return { file: f, step: match ? parseInt(match[1], 10) : 0 };
+      })
+      .sort((a, b) => b.step - a.step);
+
+    for (let i = keepCount; i < snapshots.length; i++) {
+      await unlink(join(cpDir, snapshots[i].file)).catch(() => {});
+    }
+  } catch { /* ignore */ }
+}
+
+export async function saveHealthySnapshot({ dir, runId, step, history, state, result, usage = {} }: { dir: any; runId: any; step: any; history: any; state: any; result: any; usage?: any }) {
+  const cpDir = sessionDir(dir, runId);
+  await mkdir(cpDir, { recursive: true });
+
+  const snapshot = {
+    type: 'healthy',
+    runId,
+    step,
+    history: history.map((h: any) => ({
+      step: h.step,
+      rationale: h.rationale,
+      action: h.action,
+      result: typeof h.result === 'string' ? h.result.slice(0, 2000) : '',
+      url: h.url,
+      title: h.title,
+      observation: h.observation ? { url: h.observation.url, title: h.observation.title, text: typeof h.observation.text === 'string' ? h.observation.text.slice(0, 500) : undefined } : undefined,
+    })),
+    state: sanitizeState(state),
+    usage: usage || {},
+    health: assessHealth(result),
+    timestamp: Date.now(),
+  };
+
+  const filePath = healthyPath(dir, runId, step);
+  const tmpFile = filePath + '.tmp';
+  await writeFile(tmpFile, JSON.stringify(snapshot), 'utf8');
+  await rename(tmpFile, filePath);
+
+  await pruneSnapshots(dir, runId, 'healthy', KEEP_HEALTHY);
+}
+
+export async function loadLatestHealthySnapshot(dir: string, runId: string, upToStep: number) {
+  const cpDir = sessionDir(dir, runId);
+  try {
+    const files = await readdir(cpDir);
+    const healthyFiles = files
+      .filter(f => f.startsWith('session-healthy-') && f.endsWith('.json'))
+      .map(f => {
+        const match = f.match(/session-healthy-(\d+)\.json$/);
+        return { file: f, step: match ? parseInt(match[1], 10) : 0 };
+      })
+      .filter(f => f.step <= upToStep)
+      .sort((a, b) => b.step - a.step);
+
+    if (healthyFiles.length === 0) return null;
+
+    const raw = await readFile(join(cpDir, healthyFiles[0].file), 'utf8');
+    return JSON.parse(raw);
+  } catch {
+    return null;
+  }
+}
+
+export async function listSessionCheckpoints(dir: string, runId: string) {
+  const cpDir = sessionDir(dir, runId);
+  try {
+    const files = await readdir(cpDir);
+    const result = [];
+    for (const f of files) {
+      if (!f.endsWith('.json')) continue;
+      try {
+        const raw = await readFile(join(cpDir, f), 'utf8');
+        const parsed = JSON.parse(raw);
+        result.push({
+          step: parsed.step,
+          type: parsed.type,
+          health: parsed.health,
+          timestamp: parsed.timestamp,
+          error: parsed.error || null,
+          file: f,
+        });
+      } catch { /* skip corrupt files */ }
+    }
+    return result.sort((a, b) => a.step - b.step);
+  } catch {
+    return [];
+  }
+}
+
+// ─── 导出常量 ────────────────────────────────────────────
+
+export { HEALTH_CHECKPOINT_INTERVAL, KEEP_HEALTHY, KEEP_FAILED };
